Support module graph prune and lazy load.
This bumps the go directive for the modules to 1.17 which allows the new
module graph pruning and lazy loading capabilities introduced in Go 1.17
to be used.

This means that, as described by the documentation, the go.mod file for
the module now includes a separate require block that includes all of
the indirect dependencies.